added interval to update time row based on current time
